generate an abstract class for PolyDataObject, the specific PolyDataObject class keep the textures properties, the polydataabstract is more general and will be used for tractogram with vtk polydata
